% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{ovbar_meta}
\alias{ovbar_meta}
\title{Returns a list of additional meta data of directories found in the vba
binary. This is probably not useful for the user.
The list is sorted alphabetically.}
\usage{
ovbar_meta(name)
}
\arguments{
\item{name}{the path to the input file}
}
\description{
Returns a list of additional meta data of directories found in the vba
binary. This is probably not useful for the user.
The list is sorted alphabetically.
}
