% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{ovbar_out}
\alias{ovbar_out}
\title{Returns a list of macro sources found in the vba binary.
The list is sorted alphabetically.}
\usage{
ovbar_out(name)
}
\arguments{
\item{name}{the path to the input file}
}
\description{
Returns a list of macro sources found in the vba binary.
The list is sorted alphabetically.
}
